files_lru_cache: treat LRU cache with hardlink correctly
previously, the local storage cache was based on the filenames. it
ignored, whether some of its files have the same inode, i.e. whether
some of its files are hardlinked

in cache_fn, and cache_call I use hardlinks within the local
storage. Files that are hardlinked should not be accounted
twice.

Although, this does not really seem to be a problem in pvgrip. There,
only one of the hardlinks is heavily utilised, and hence another one
is remove in time. However, in general I decided to address this
problem for `cu`.

The solution consists of:
  - mapping: inode -> size (previously was filename -> size)
  - introducing mapping: filename -> inode
  - introducing mapping: inode -> number of links

`_add_fn` method reviews, whether filename's inode has been changed,
and whether the corresponding inode size has been changed.

Extra tests that test new functionality are added.